Refactor: new neuromorphic package, tests, Makefiles, and CI workflow
- Replace flat SNNTraining/ with installable neuromorphic/ package
  - config.py  (ModelConfig/HardwareSpecs, compat aliases Variables/Specs)
  - model.py   (SpikingNet, Metrics – removed debug prints, fixed return unpack)
  - dataset.py (NavDataset – type hints, removed dead example code)
  - trainer.py (Trainer – fixed `output, _ = net()` bug, logging, clean pruning)
  - utils.py   (no global state: attach_hooks returns spike_record, counter class)
  - graph.py   (Graph – logging, type hints)
  - mapping.py (Mapping – type hints, clear docstring)
  - packets.py (init_pipeline, record_packets, DynamicInference)
  - sweep.py   (SweepHandler)

- Add scripts/ entry points (argparse, sys.path safe)
  train.py · export_nir.py · evaluate_activity.py · evaluate_results.py · map_neurons.py

- Add tests/ suite (pytest, no GPU required, 5 test modules)
  test_config · test_dataset · test_model · test_mapping · test_utils

- Add pyproject.toml (setuptools, dev/sweep/analysis extras)
- Add root Makefile (install test lint build-cpp train export-nir evaluate-* simulate clean)
- Add RoutingEval/Makefile delegating to common/makefile
- Add .github/workflows/ci.yml (python, cpp, smoke jobs)
- Fix C++ -Wreorder warnings by reordering header member declarations
- Fix C++ -Wsign-compare in RoutingSimulator.cpp (upPath loop)

+++ RoutingEval/common/RoutingSimulator.cpp
@@ -149,7 +149,7 @@ void RoutingSimulator::simulate() {
         std::unordered_set<int> visited;
 
         int i = 0;
-        while (i < upPath.size() && upPath[i] != rootCore && upPath[i] != lcaCore) {
+        while (i < static_cast<int>(upPath.size()) && upPath[i] != rootCore && upPath[i] != lcaCore) {
             routingPath += "U";
             visited.insert(upPath[i]);
             ++i;
@@ -159,7 +159,7 @@ void RoutingSimulator::simulate() {
             routingPath += "B";  // reached LCA directly
         } else {
             routingPath += "D";  // start descent to LCA
-            for (++i; i < upPath.size(); ++i) {
+            for (++i; i < static_cast<int>(upPath.size()); ++i) {
                 int parent = upPath[i - 1];
                 int child = upPath[i];
                 routingPath += (child < parent ? "L" : "R");
